Put the test in a different location
